requirements.txt
# Crop vs. Non-Crop Land Segmentation Requirements
# CS30305 - Foundations of Data Science Core
# Student: Her Saleh 22 BCE 10564

# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Geospatial and Image Processing
gdal>=3.2.0
rasterio>=1.2.0
geopandas>=0.10.0
opencv-python>=4.5.0
scikit-image>=0.18.0
rasterstats>=0.15.0
xarray>=0.19.0

# Deep Learning Frameworks
tensorflow>=2.8.0
keras>=2.8.0
# Alternative: torch>=1.10.0 torchvision>=0.11.0

# Visualization and Plotting
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and Development
jupyterlab>=3.2.0
notebook>=6.4.0
ipywidgets>=7.6.0

# Utility Libraries
tqdm>=4.62.0
pyyaml>=5.4.0
pillow>=8.3.0
h5py>=3.1.0

# Statistical Analysis
statsmodels>=0.12.0
pysal>=2.4.0

# Web Scraping and Data Collection
beautifulsoup4>=4.10.0
requests>=2.26.0

# Cloud and Storage
boto3>=1.18.0

# Performance and Optimization
cython>=0.29.0
numba>=0.54.0

# Testing and Quality
pytest>=6.2.0
pylint>=2.10.0
line_profiler>=3.3.0

# Documentation
sphinx>=4.2.0

# Additional Utilities
future>=0.18.0
graphviz>=0.17.0
sqlalchemy>=1.4.0
nodejs>=16.0.0

# Optional: GPU Support (uncomment if using CUDA)
# tensorflow-gpu>=2.8.0
# torch>=1.10.0+cu113 --extra-index-url https://download.pytorch.org/whl/cu113

# Development Dependencies
black>=21.9.0
flake8>=3.9.0
isort>=5.9.0